company-loop/requirements.txt
# Core Runtime Dependencies
python-dotenv>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0

# API Clients
openai>=1.10.0
anthropic>=0.8.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# ClickUp (REST API)
requests>=2.31.0

# Close CRM (REST API)
closeio>=0.1.0

# Social Media APIs
google-api-python-client>=2.100.0  # YouTube
tweepy>=4.14.0  # Twitter/X
# TikTok, Instagram, LinkedIn use requests with OAuth

# Date/Time
python-dateutil>=2.8.2
pytz>=2023.3

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Async Support
asyncio>=3.4.3
aiohttp>=3.9.0

# Logging & Monitoring
loguru>=0.7.0
sentry-sdk>=1.39.0  # Optional: error tracking

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code Quality
black>=23.7.0
ruff>=0.0.292
mypy>=1.5.0

# Utilities
humanize>=4.8.0
tenacity>=8.2.3  # Retry logic
jsonschema>=4.19.0
